+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * (DatePipe): new date pipe. It is due to Angular's date pipe has format issue on IE11;
 * (noop): noop utils function;
 * (Modal): Modal animation feature toggle;
+* (RequiredWithTrim): Required validation for trim value;
 
 # 0.0.21 (2017-08-28)